+++ pylint_pytest/checkers/fixture.py
@@ -5,7 +5,7 @@ import pylint
 from pylint.checkers.variables import VariablesChecker
 from pylint.interfaces import IAstroidChecker
 import pytest
-from ..utils import _can_use_fixture, _is_same_module, _is_pytest_mark_usefixtures
+from ..utils import _can_use_fixture, _is_same_module, _is_pytest_mark_usefixtures, _is_pytest_fixture
 from . import BasePytestChecker
 
 
@@ -17,9 +17,15 @@ class FixtureCollector:
         self.fixtures = session._fixturemanager._arg2fixturedefs
 
 
-class FixtureLoader(BasePytestChecker):
+class FixtureChecker(BasePytestChecker):
     __implements__ = IAstroidChecker
-    msgs = {'E6401': ('', 'pytest-fixture-loader', '')}
+    msgs = {
+        'W6401': (
+            'Using a deprecated @pytest.yield_fixture decorator',
+            'deprecated-pytest-yield-fixture',
+            'Used when using a deprecated pytest decorator that has been deprecated in pytest-3.0'
+        ),
+    }
 
     _pytest_fixtures = {}
     _invoked_with_func_args = set()
@@ -28,19 +34,19 @@ class FixtureLoader(BasePytestChecker):
 
     def open(self):
         # patch VariablesChecker.add_message
-        FixtureLoader._original_add_message = VariablesChecker.add_message
-        VariablesChecker.add_message = FixtureLoader.patch_add_message
+        FixtureChecker._original_add_message = VariablesChecker.add_message
+        VariablesChecker.add_message = FixtureChecker.patch_add_message
 
     def close(self):
         '''restore & reset class attr for testing'''
         # restore add_message
-        VariablesChecker.add_message = FixtureLoader._original_add_message
-        FixtureLoader._original_add_message = callable
+        VariablesChecker.add_message = FixtureChecker._original_add_message
+        FixtureChecker._original_add_message = callable
 
         # reset fixture info storage
-        FixtureLoader._pytest_fixtures = {}
-        FixtureLoader._invoked_with_func_args = set()
-        FixtureLoader._invoked_with_usefixtures = set()
+        FixtureChecker._pytest_fixtures = {}
+        FixtureChecker._invoked_with_func_args = set()
+        FixtureChecker._invoked_with_usefixtures = set()
 
     def visit_module(self, node):
         '''
@@ -49,13 +55,13 @@ class FixtureLoader(BasePytestChecker):
         - create containers for the module to store args and fixtures
         '''
         # storing all fixtures discovered by pytest session
-        FixtureLoader._pytest_fixtures = {}  # Dict[List[_pytest.fixtures.FixtureDef]]
+        FixtureChecker._pytest_fixtures = {}  # Dict[List[_pytest.fixtures.FixtureDef]]
 
         # storing all used function arguments
-        FixtureLoader._invoked_with_func_args = set()  # Set[str]
+        FixtureChecker._invoked_with_func_args = set()  # Set[str]
 
         # storing all invoked fixtures through @pytest.mark.usefixture(...)
-        FixtureLoader._invoked_with_usefixtures = set()  # Set[str]
+        FixtureChecker._invoked_with_usefixtures = set()  # Set[str]
 
         try:
             with open(os.devnull, 'w') as devnull:
@@ -69,7 +75,7 @@ class FixtureLoader(BasePytestChecker):
                     [node.file, '--fixtures'],
                     plugins=[fixture_collector],
                 )
-                FixtureLoader._pytest_fixtures = fixture_collector.fixtures
+                FixtureChecker._pytest_fixtures = fixture_collector.fixtures
         finally:
             # restore output devices
             sys.stdout, sys.stderr = stdout, stderr
@@ -87,6 +93,10 @@ class FixtureLoader(BasePytestChecker):
                         # save all visited fixtures
                         for arg in decorator.args:
                             self._invoked_with_usefixtures.add(arg.value)
+                    if int(pytest.__version__.split('.')[0]) >= 3 and \
+                            _is_pytest_fixture(decorator, fixture=False):
+                        # raise deprecated warning for @pytest.yield_fixture
+                        self.add_message('deprecated-pytest-yield-fixture', node=node)
             for arg in node.args.args:
                 self._invoked_with_func_args.add(arg.name)
 
@@ -109,17 +119,17 @@ class FixtureLoader(BasePytestChecker):
                 pass
 
             # imported fixture is referenced in test/fixture func
-            elif fixture_name in FixtureLoader._invoked_with_func_args \
-                    and fixture_name in FixtureLoader._pytest_fixtures:
-                if _is_same_module(fixtures=FixtureLoader._pytest_fixtures,
+            elif fixture_name in FixtureChecker._invoked_with_func_args \
+                    and fixture_name in FixtureChecker._pytest_fixtures:
+                if _is_same_module(fixtures=FixtureChecker._pytest_fixtures,
                                    import_node=node,
                                    fixture_name=fixture_name):
                     return
 
             # fixture is referenced in @pytest.mark.usefixtures
-            elif fixture_name in FixtureLoader._invoked_with_usefixtures \
-                    and fixture_name in FixtureLoader._pytest_fixtures:
-                if _is_same_module(fixtures=FixtureLoader._pytest_fixtures,
+            elif fixture_name in FixtureChecker._invoked_with_usefixtures \
+                    and fixture_name in FixtureChecker._pytest_fixtures:
+                if _is_same_module(fixtures=FixtureChecker._pytest_fixtures,
                                    import_node=node,
                                    fixture_name=fixture_name):
                     return
@@ -128,20 +138,20 @@ class FixtureLoader(BasePytestChecker):
         if msgid == 'unused-argument' and \
                 _can_use_fixture(node.parent.parent) and \
                 isinstance(node.parent, astroid.Arguments) and \
-                node.name in FixtureLoader._pytest_fixtures:
+                node.name in FixtureChecker._pytest_fixtures:
             return
 
         # check W0621 redefined-outer-name
         if msgid == 'redefined-outer-name' and \
                 _can_use_fixture(node.parent.parent) and \
                 isinstance(node.parent, astroid.Arguments) and \
-                node.name in FixtureLoader._pytest_fixtures:
+                node.name in FixtureChecker._pytest_fixtures:
             return
 
         if int(pylint.__version__.split('.')[0]) >= 2:
-            FixtureLoader._original_add_message(
+            FixtureChecker._original_add_message(
                 self, msgid, line, node, args, confidence, col_offset)
         else:
             # python2 + pylint1.9 backward compatibility
-            FixtureLoader._original_add_message(
+            FixtureChecker._original_add_message(
                 self, msgid, line, node, args, confidence)