+++ tests/input/deprecated-pytest-yield-fixture/func.py
@@ -1,0 +1,11 @@
+import pytest
+
+
+@pytest.yield_fixture()
+def yield_fixture():
+    yield
+
+
+@pytest.yield_fixture(scope='session')
+def yield_fixture_session():
+    yield