+++ tests/input/deprecated-pytest-yield-fixture/smoke.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.yield_fixture
+def yield_fixture():
+    yield